requirements.txt
# Core LLM Provider SDKs
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.0.0
groq>=0.4.0
ollama>=0.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0

# Development (optional)
pytest>=7.0.0
black>=23.0.0
mypy>=1.0.0

# Notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
